+++ leaves/models.py
@@ -3,4 +3,4 @@ from django.db import models
 
 class Leave(models.Model):
     name = models.CharField(verbose_name="휴가 종류", max_length=30)
-    consume = models.IntegerField(verbose_name="소진일")
+    consume = models.FloatField(verbose_name="소진일")
